+++ CHANGELOG.md
@@ -3,5 +3,9 @@
 ##  Unreleased
 * Placeholder
 
-## [0.6.8] 2021-3-6 
+## [0.6.9] 2021-7-17
+* Support a new interface, `raw_header()` to add raw plantuml text as part of generated diagram.
+* Support a new command line option, `raw_header_file` for the same purpose.
+
+## [0.6.8] 2021-3-6
 * Support a new interface, `outside()` to express the call coming from outside of the diagram.